+++ extensions/safari/src/lang-adaptor/index.ts
@@ -1,0 +1,2 @@
+export * from "./pinyin";
+export { LanguageAdaptor } from "./base";