+++ extensions/safari/src/utils.ts
@@ -3,11 +3,13 @@ import Fuse, { FuseOptionKey } from "fuse.js";
 import _ from "lodash";
 import osascript from "osascript-tag";
 import { URL } from "url";
-import { HistoryItem, Tab } from "./types";
+import { pinyinSupport, LanguageAdaptor } from './lang-adaptor';
+import { HistoryItem, LooseTab, Tab } from "./types";
 
 type Preferences = {
   safariAppIdentifier: string;
 };
+
 
 export const { safariAppIdentifier }: Preferences = getPreferenceValues();
 
@@ -71,12 +73,40 @@ export const getTitle = (tab: Tab) => _.truncate(tab.title, { length: 75 });
 
 export const plural = (count: number, string: string) => `${count} ${string}${count > 1 ? "s" : ""}`;
 
-export const search = function (collection: object[], keys: Array<FuseOptionKey<object>>, searchText: string) {
+const langAdaptor = new LanguageAdaptor();
+const langHandlers = [pinyinSupport()]
+
+export const search = function (collection: LooseTab[], keys: Array<FuseOptionKey<object>>, searchText: string) {
+  langHandlers.forEach((handler) => {
+    handler.install(langAdaptor);
+  });
+
   if (!searchText) {
     return collection;
   }
 
-  return new Fuse(collection, { keys, threshold: 0.35 }).search(searchText).map((x) => x.item);
+
+  const _formatPerf = performance.now();
+  const formattedCollection = collection.map((item) => {
+    return {
+      ...item,
+      title_formatted: langAdaptor.formatString(searchText, item.title, { id: item.uuid })
+    }
+  });
+  const _formatCost = performance.now() - _formatPerf;
+
+  const _searchPerf = performance.now();
+  const result = new Fuse(formattedCollection, { keys, threshold: 0.35 }).search(searchText).map((x) => x.item);
+  const _searchCost = performance.now() - _searchPerf;
+
+  if (process.env.NODE_ENV === 'development') {
+    console.log('searchText', searchText);
+    console.log(`format cost ${_formatCost}ms`);
+    console.log(`search cost ${_searchCost}ms`);
+    // console.log('formatted collection', formattedCollection);
+    console.log('result size', result.length);
+  }
+  return result;
 };
 
 const dtf = new Intl.DateTimeFormat(undefined, {